+++ attributes/default.rb
@@ -3,6 +3,7 @@ default attributes for aq installation cookbook
 =end
 
 default['aq_config']['enable_https'] = false
+default['aq_config']['xsl_report_url'] = 'ftp://ftpext.usgs.gov/pub/cr/la/ruston/Blob%20templates/WebApps.ServiceModel.ResponseDtos.Visits.VisitInfo.xsl'
 default['aq_config']['database_databag'] = 'migration_db'
 default['aq_config']['databag'] = 'aquarius-oregon-aquarius-development'
 default['aq_config']['license_databag'] = 'aq_license_caws_dev'